api/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
httpx>=0.27.0
pydantic>=2.0.0
slowapi>=0.1.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
